Resolve Issue #15 and #93
